Added modules for Dataset for pure substances and binary mixtures
