--- crates/feos-core/src/ad/parameter_optimization/dataset/mod.rs
@@ -1,210 +1,0 @@
-mod binary;
-mod pure;
-
-use std::{io, path::Path, sync::Arc};
-
-use ndarray::{Array1, Array2, ArrayView1, ArrayView2};
-use serde::Serialize;
-use serde::de::DeserializeOwned;
-
-use crate::{ParametersAD, Residual};
-
-pub use binary::{BinaryDataset, BinaryProperty};
-pub use pure::{PureDataset, PureProperty};
-
-/// Per-dataset evaluation result: inputs, experimental target, model
-/// prediction, and derived statistics at a given set of parameters.
-///
-/// Produced by [`crate::Regressor::evaluate_datasets`].
-/// Fully serializable and convertable into any tabular format (CSV, JSON, DataFrame).
-#[derive(Debug, Serialize)]
-pub struct DatasetResult {
-    /// Dataset name (default property name or user-supplied).
-    pub name: String,
-    /// Reported independent input column names and their values.
-    pub inputs: Vec<(&'static str, Vec<f64>)>,
-    /// Name of the target property column.
-    pub target_name: &'static str,
-    /// Experimental target values.
-    pub target: Vec<f64>,
-    /// Values at the given parameters predicted by the model.
-    /// `NaN` for points where calculations did not converge.
-    pub predicted: Vec<f64>,
-    /// Whether the calculation converged for each point.
-    pub converged: Vec<bool>,
-    /// Relative deviation `(predicted − target) / target`.
-    /// `NaN` for non-converged points.
-    pub relative_deviation: Vec<f64>,
-}
-
-/// Shared numerical data for all datasets.
-struct DatasetData {
-    inputs: Array2<f64>,
-    target: Array1<f64>,
-}
-
-/// Shared in-memory representation for all datasets.
-///
-/// Cheap to clone: the inputs/target arrays are kept behind an `Arc`,
-/// while the user-supplied name is owned per handle.
-#[derive(Clone)]
-struct DatasetStorage {
-    data: Arc<DatasetData>,
-    name: Option<String>,
-}
-
-impl DatasetStorage {
-    fn from_records<R: DatasetRecord>(records: Vec<R>) -> Self {
-        let n = records.len();
-        let inputs = Array2::from_shape_fn((n, R::N_INPUTS), |(i, j)| records[i].input(j));
-        let target = Array1::from_iter(records.iter().map(DatasetRecord::target));
-        Self {
-            data: Arc::new(DatasetData { inputs, target }),
-            name: None,
-        }
-    }
-
-    fn from_csv<R: DatasetRecord>(path: &Path) -> Result<Self, csv::Error> {
-        let records = csv::Reader::from_path(path)?
-            .deserialize()
-            .collect::<Result<Vec<R>, _>>()?;
-        Ok(Self::from_records(records))
-    }
-
-    fn from_reader<R: DatasetRecord>(reader: impl io::Read) -> Result<Self, csv::Error> {
-        let records = csv::Reader::from_reader(reader)
-            .deserialize()
-            .collect::<Result<Vec<R>, _>>()?;
-        Ok(Self::from_records(records))
-    }
-
-    fn inputs(&self) -> ArrayView2<'_, f64> {
-        self.data.inputs.view()
-    }
-
-    fn target(&self) -> ArrayView1<'_, f64> {
-        self.data.target.view()
-    }
-
-    fn name(&self) -> Option<&str> {
-        self.name.as_deref()
-    }
-
-    fn set_name(&mut self, name: String) {
-        self.name = Some(name);
-    }
-}
-
-/// Conversion logic for records that can be collected into a dataset.
-pub trait DatasetRecord: DeserializeOwned {
-    /// Number of columns passed to the property evaluator.
-    const N_INPUTS: usize;
-
-    /// Value of model input column `column`.
-    fn input(&self, column: usize) -> f64;
-
-    /// Experimental target value.
-    fn target(&self) -> f64;
-}
-
-/// Experimental data container with metadata and a non-AD model evaluator.
-///
-/// Implementors expose the inputs and targets plus an
-/// [`evaluate`](Self::evaluate) method that runs the dataset's property
-/// against a single model without computing gradients. Used for model
-/// comparison and CLI/Python multi-model workflows.
-pub trait Dataset {
-    /// Inputs for model evaluation, shape `[n_points, k]`.
-    fn inputs(&self) -> ArrayView2<'_, f64>;
-
-    /// Target values, shape `[n_points]`.
-    fn target(&self) -> ArrayView1<'_, f64>;
-
-    /// Property name used for logging and diagnostics.
-    fn name(&self) -> &str;
-
-    /// Names of independent input columns reported in diagnostics.
-    ///
-    /// These can be fewer than the number of columns in [`Self::inputs`] when
-    /// the target is also passed to the model as an initial guess.
-    fn input_names(&self) -> &'static [&'static str];
-
-    /// Name of the target property column.
-    fn target_name(&self) -> &'static str;
-
-    /// Evaluate this dataset's property against a single model.
-    ///
-    /// Returns `(predicted, converged)`:
-    /// - `predicted`: shape `[n_points]`, in SI units; `NaN` where the
-    ///   underlying solver did not converge.
-    /// - `converged`: shape `[n_points]`.
-    fn evaluate<E>(&self, model: &E) -> (Array1<f64>, Array1<bool>)
-    where
-        E: Residual + Sync;
-}
-
-/// Maximum number of parameters that can be fitted simultaneously.
-///
-/// Enforced upstream at `Regressor::new`. The match arms generated by
-/// [`impl_evaluate_ad!`] must cover `1..=MAX_FITTED_PARAMETERS`; bumping
-/// this constant means widening that list too.
-pub(crate) const MAX_FITTED_PARAMETERS: usize = 14;
-
-/// Emit the default body of [`DatasetAD::evaluate_ad`]: a runtime-`P` →
-/// const-`P` dispatch covering each listed parameter count.
-macro_rules! impl_evaluate_ad {
-    ($($p:literal),+ $(,)?) => {
-        /// Evaluate the property and its parameter gradients at the given parameters.
-        ///
-        /// - `param_names`: names of the `P` parameters being differentiated.
-        /// - `params`: the full parameter vector; only entries listed in `param_names` are seeded.
-        ///
-        /// Returns `(predicted, gradients, converged)`:
-        /// - `predicted`: shape `[n_points]`, in SI units.
-        /// - `gradients`: shape `[n_points, P]`.
-        /// - `converged`: shape `[n_points]`.
-        fn evaluate_ad<T: ParametersAD<N>>(
-            &self,
-            param_names: &[String],
-            params: &[f64],
-        ) -> (Array1<f64>, Array2<f64>, Array1<bool>) {
-            let n = self.inputs().nrows();
-            let parameters = Array2::from_shape_fn((n, params.len()), |(_, j)| params[j]);
-
-            fn to_const<const P: usize>(names: &[String]) -> [String; P] {
-                names.to_vec().try_into().expect("parameter count mismatch")
-            }
-
-            match param_names.len() {
-                $(
-                    $p => self.evaluate_ad_const::<T, $p>(
-                        to_const(param_names),
-                        parameters.view(),
-                        self.inputs(),
-                    ),
-                )+
-                p => unreachable!(
-                    "Regressor::new rejects fit lists longer than \
-                     MAX_FITTED_PARAMETERS={MAX_FITTED_PARAMETERS}; got {p}",
-                ),
-            }
-        }
-    };
-}
-
-/// Experimental data container that supports parameter-gradient evaluation
-/// for models implementing [`ParametersAD<N>`].
-pub trait DatasetAD<const N: usize>: Dataset {
-    /// Evaluate the property and its `P` parameter gradients at compile-time-known `P`.
-    ///
-    /// Implementors typically delegate to a property-specific
-    /// `*_parallel_ad::<T, P>` function.
-    fn evaluate_ad_const<T: ParametersAD<N>, const P: usize>(
-        &self,
-        names: [String; P],
-        parameters: ArrayView2<f64>,
-        inputs: ArrayView2<f64>,
-    ) -> (Array1<f64>, Array2<f64>, Array1<bool>);
-
-    impl_evaluate_ad!(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14);
-}
