--- crates/feos-core/src/ad/parameter_optimization/mod.rs
@@ -1,11 +1,0 @@
-pub mod dataset;
-
-pub use crate::ad::properties::{
-    BubblePointRecord, DewPointRecord, EnthalpyOfVaporizationRecord,
-    EquilibriumLiquidDensityRecord, LiquidDensityRecord, ResidualIsobaricHeatCapacityRecord,
-    VaporPressureRecord,
-};
-pub use dataset::{
-    BinaryDataset, BinaryProperty, Dataset, DatasetAD, DatasetRecord, DatasetResult, PureDataset,
-    PureProperty,
-};
