+++ crates/feos-core/src/ad/properties/dew_point_pressure.rs
@@ -1,5 +1,5 @@
 use crate::ad::Gradient;
-use crate::ad::parameter_optimization::dataset::DatasetRecord;
+use crate::ad::dataset::DatasetRecord;
 use crate::ad::{ParametersAD, vectorize, vectorize_ad};
 use crate::{Composition, Contributions, FeosResult, PhaseEquilibrium, ReferenceSystem, Residual};
 use nalgebra::{SVector, U2};
