+++ crates/feos-core/src/ad/properties/mod.rs
@@ -7,37 +7,11 @@ pub mod liquid_density;
 pub mod residual_isobaric_heat_capacity;
 pub mod vapor_pressure;
 
-pub use boiling_temperature::{boiling_temperature, boiling_temperature_ad};
-pub use bubble_point_pressure::{
-    BubblePointRecord, bubble_point_pressure, bubble_point_pressure_ad,
-};
-pub use dew_point_pressure::{DewPointRecord, dew_point_pressure, dew_point_pressure_ad};
-pub use enthalpy_of_vaporization::{
-    EnthalpyOfVaporizationRecord, enthalpy_of_vaporization, enthalpy_of_vaporization_ad,
-};
-pub use equilibrium_liquid_density::{
-    EquilibriumLiquidDensityRecord, equilibrium_liquid_density, equilibrium_liquid_density_ad,
-};
-pub use liquid_density::{LiquidDensityRecord, liquid_density, liquid_density_ad};
-pub use residual_isobaric_heat_capacity::{
-    ResidualIsobaricHeatCapacityRecord, residual_isobaric_heat_capacity,
-    residual_isobaric_heat_capacity_ad,
-};
-pub use vapor_pressure::{VaporPressureRecord, vapor_pressure, vapor_pressure_ad};
-
-pub use boiling_temperature::{boiling_temperature_parallel, boiling_temperature_parallel_ad};
-pub use bubble_point_pressure::{
-    bubble_point_pressure_parallel, bubble_point_pressure_parallel_ad,
-};
-pub use dew_point_pressure::{dew_point_pressure_parallel, dew_point_pressure_parallel_ad};
-pub use enthalpy_of_vaporization::{
-    enthalpy_of_vaporization_parallel, enthalpy_of_vaporization_parallel_ad,
-};
-pub use equilibrium_liquid_density::{
-    equilibrium_liquid_density_parallel, equilibrium_liquid_density_parallel_ad,
-};
-pub use liquid_density::{liquid_density_parallel, liquid_density_parallel_ad};
-pub use residual_isobaric_heat_capacity::{
-    residual_isobaric_heat_capacity_parallel, residual_isobaric_heat_capacity_parallel_ad,
-};
-pub use vapor_pressure::{vapor_pressure_parallel, vapor_pressure_parallel_ad};
+pub use boiling_temperature::*;
+pub use bubble_point_pressure::*;
+pub use dew_point_pressure::*;
+pub use enthalpy_of_vaporization::*;
+pub use equilibrium_liquid_density::*;
+pub use liquid_density::*;
+pub use residual_isobaric_heat_capacity::*;
+pub use vapor_pressure::*;
